Safari-611.1.10/LayoutTests/platform/mac/fast/overflow/overflow-scrollbar-hit-test-expected.txt

Inner

Inner
Testing clicks on box1
Scrolled down by 80 on down arrow click: PASS
Scrolled up by 40 on up arrow click: PASS

Testing clicks on box2
Scrolled down by 80 on down arrow click: PASS
Scrolled up by 40 on up arrow click: PASS
